--refactor: imgUpload and payment to services
