+++ services/payment.js
@@ -1,9 +1,9 @@
 const razorpay = require("razorpay");
-require("dotenv").config();
+const { RZP_KEY_ID, RZP_KEY_SECRET } = require("../config/index");
 
 let rzp = new razorpay({
-	key_id: process.env.RZP_KEY_ID,
-	key_secret: process.env.RZP_KEY_SECRET
+	key_id: RZP_KEY_ID,
+	key_secret: RZP_KEY_SECRET
 });
 
 module.exports.createRzpOrder = async (amt, email) => {
